pyproject.toml
[project]
name = "coltec-codespaces"
version = "0.1.0"
description = "Coltec devcontainer specs and tooling"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.6",
    "pydantic>=2.7",
    "pyyaml>=6.0",
]

[project.scripts]
coltec-codespaces = "coltec_codespaces.__main__:main"
